pyproject.toml
[tool.poetry]
name = "finvestor"
version = "0.1.0"
description = "Code for tracking and making sense of your different investement platforms."
authors = ["Ouail Bendidi <ouail.bendidi@gmail.com>"]
license = "MIT"
packages = [
    {include = "finvestor"}
]

[tool.poetry.scripts]
finvestor = "finvestor.cli:app"

[tool.poetry.dependencies]
python = "^3.8"
toml = "^0.10.2"
pandas = "^1.3.4"
openpyxl = "^3.0.9"
pydantic = "^1.8.2"
typer = "^0.4.0"
jupyter = {version = "^1.0.0", optional = true}
jupyter-contrib-nbextensions = {version = "^0.5.1", optional = true}
attrs = "^21.2.0"
alpaca-trade-api = "^1.4.1"
httpx = "^0.20.0"
lxml = "^4.6.4"
trio = "^0.19.0"

[tool.poetry.dev-dependencies]
black = {version = "^21.9b0", allow-prereleases = true}
mypy = "^0.910"
isort = "^5.9.3"
flake8 = "^4.0.1"
pytest = "^6.2.5"
pytest-cov = "^3.0.0"
syrupy = "^1.5.0"
pandas-stubs = "^1.2.0"
pre-commit = "^2.15.0"
types-pytz = "^2021.3.0"
respx = "^0.18.2"

[tool.poetry.extras]
jupyter = ["jupyter", "jupyter-contrib-nbextensions"]

[build-system]
requires = ["poetry-core>=1.10.0"]
build-backend = "poetry.core.masonry.api"

[mypy]
python_version = 3.8
plugins = "pydantic.mypy"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
known_first_party = ["finvestor"]

[tool.pytest.ini_options]
addopts = "-ra -q -vv --cov=finvestor --cov-report=term --cov-report=xml"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
branch = true

[tool.coverage.report]
show_missing = true
omit = ["**/tests/test_*.py"]
